# Conflicts:
#	nJMaLBot/Utilities/Commands/CommandHandler.cs
#	nJMaLBot/nJMaLBot.csproj

+++ nJMaLBot/Commands/Music.cs
@@ -12,8 +12,10 @@ using Bot.Utilities.Commands;
 using Bot.Utilities.Modules;
 using Discord;
 using Discord.Commands;
+using Lavalink4NET.Player;
 using Lavalink4NET.Rest;
 using LiteDB;
+using Tyrrrz.Extensions;
 
 #pragma warning disable 4014
 
@@ -28,9 +30,9 @@ namespace Bot.Commands {
         public async Task Play([Remainder] [Summary("play0_0s")] string query = null) {
             if (!await IsPreconditionsValid)
                 return;
-            var logMessage = GetLogMessage();
 
-            Player.SetControlMessage(await logMessage);
+            Player.EnqueueControlMessageSend(ResponseChannel);
+            
             try {
                 await MusicUtils.QueueLoadMusic(Context.Message, query, Player);
             }
@@ -42,7 +44,7 @@ namespace Bot.Commands {
                 if (Player.Playlist.Count == 0) Player.ControlMessage.SafeDelete();
             }
             catch (AttachmentAddFailException) {
-                ReplyFormattedAsync(Loc.Get("Music.AttachmentFail"), true, await logMessage).DelayedDelete(TimeSpan.FromMinutes(10));
+                ReplyFormattedAsync(Loc.Get("Music.AttachmentFail"), true).DelayedDelete(TimeSpan.FromMinutes(10));
                 if (Player.Playlist.Count == 0) Player.ControlMessage.SafeDelete();
             }
         }
@@ -52,13 +54,9 @@ namespace Bot.Commands {
         [Summary("stop0s")]
         public async Task Stop() {
             if (!await IsPreconditionsValid) return;
-            if (Player?.CurrentTrack == null) {
-                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(2));
-                return;
-            }
 
-            Player.PrepareShutdown(Loc.Get("Music.UserStopPlayback").Format(Context.User.Username));
-            await Player.StopAsync(true);
+            Player.Dispose(Loc.Get("Music.UserStopPlayback").Format(Context.User.Username), false);
+            EmbedPlaybackControl.ForceRemove(Context.Guild.Id, Loc.Get("Music.UserStopPlayback").Format(Context.User.Username), false);
         }
 
         [Command("jump", RunMode = RunMode.Async)]
@@ -66,8 +64,8 @@ namespace Bot.Commands {
         [Summary("jump0s")]
         public async Task Jump([Summary("jump0_0s")] int index = 1) {
             if (!await IsPreconditionsValid) return;
-            if (Player == null) {
-                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(2));
+            if (Player == null || Player.Playlist.IsEmpty) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
                 return;
             }
 
@@ -82,7 +80,11 @@ namespace Bot.Commands {
         [Summary("goto0s")]
         public async Task Goto([Summary("goto0_0s")] int index) {
             if (!await IsPreconditionsValid) return;
-            
+            if (Player == null || Player.Playlist.IsEmpty) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
+                return;
+            }
+
             //For programmers who count from 0
             if (index == 0) index = 1;
 
@@ -104,7 +106,7 @@ namespace Bot.Commands {
         public async Task Volume([Summary("volume0_0s")] int volume = 100) {
             if (!await IsPreconditionsValid) return;
             if (Player == null) {
-                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(2));
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
                 return;
             }
 
@@ -123,7 +125,7 @@ namespace Bot.Commands {
         public async Task Repeat(LoopingState state) {
             if (!await IsPreconditionsValid) return;
             if (Player == null) {
-                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(2));
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
                 return;
             }
 
@@ -144,7 +146,7 @@ namespace Bot.Commands {
         public async Task Pause() {
             if (!await IsPreconditionsValid) return;
             if (Player == null) {
-                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(2));
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
                 return;
             }
 
@@ -158,7 +160,7 @@ namespace Bot.Commands {
         public async Task Resume() {
             if (!await IsPreconditionsValid) return;
             if (Player == null) {
-                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(2));
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
                 return;
             }
 
@@ -171,8 +173,8 @@ namespace Bot.Commands {
         [Summary("shuffle0s")]
         public async Task Shuffle() {
             if (!await IsPreconditionsValid) return;
-            if (Player == null) {
-                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(2));
+            if (Player == null || Player.Playlist.IsEmpty) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
                 return;
             }
 
@@ -191,30 +193,7 @@ namespace Bot.Commands {
                 return;
             }
 
-            var queue = new StringBuilder("```py\n");
-            for (var index = 0; index < Player.Playlist.Count; index++) {
-                var builder = new StringBuilder();
-                var lavalinkTrack = Player.Playlist[index];
-                builder.Append(Player.CurrentTrackIndex == index ? "@" : " ");
-                builder.Append(index + 1);
-                builder.Append(": ");
-                builder.AppendLine(lavalinkTrack.Title);
-                if (queue.Length + builder.Length > 2000) {
-                    await PrintList(queue, logMessage);
-                    logMessage = null;
-                    queue = new StringBuilder("```py\n");
-                }
-
-                queue.Append(builder);
-            }
-
-            await PrintList(queue, logMessage);
-
-            async Task PrintList(StringBuilder builder, IUserMessage message = null) {
-                builder.Append("```");
-                MusicUtils.EscapeTrack(builder);
-                await ReplyFormattedAsync(builder.ToString(), false, message);
-            }
+            Player.PrintQueue(logMessage);
         }
 
         [Hidden]
@@ -223,12 +202,12 @@ namespace Bot.Commands {
         [Summary("saveplaylist0s")]
         public async Task SavePlaylist() {
             if (!await IsPreconditionsValid) return;
-            if (Player == null || Player.Playlist.Count == 0) {
-                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(2));
+            if (Player == null || Player.Playlist.IsEmpty) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
                 return;
             }
 
-            var playlist = Player.GetExportPlaylist(ExportPlaylistOptions.IgnoreTrackIndex);
+            var playlist = Player.ExportPlaylist(ExportPlaylistOptions.IgnoreTrackIndex);
             var storedPlaylist = new StoredPlaylist {
                 Tracks = playlist.Tracks, TrackIndex = playlist.TrackIndex, TrackPosition = playlist.TrackPosition,
                 Id = "u" + ObjectId.NewObjectId()
@@ -264,10 +243,6 @@ namespace Bot.Commands {
         private async Task ExecutePlaylist(string id, ImportPlaylistOptions options) {
             if (!await IsPreconditionsValid) return;
             var logMessage = await GetLogMessage();
-            if (logMessage == null || Player == null) {
-                logMessage.SafeDelete();
-                return;
-            }
 
             Player.SetControlMessage(logMessage);
             var playlist = GlobalDB.Playlists.FindById(id);
@@ -288,7 +263,7 @@ namespace Bot.Commands {
         public async Task SearchYoutube([Summary("play0_0s")] [Remainder] string query) {
             await AdvancedSearch(SearchMode.YouTube, query);
         }
-        
+
         [SummonToUser]
         [Command("soundcloud", RunMode = RunMode.Async)]
         [Alias("sc")]
@@ -318,7 +293,7 @@ namespace Bot.Commands {
             var msg = await ReplyAsync(null, false, eb.Build());
             if (!tracks.Any())
                 return;
-            
+
             var controller = CollectorsUtils.CollectReaction(msg, reaction => true, async args => {
                 args.RemoveReason();
                 var i = args.Reaction.Emote.Name switch {
@@ -337,7 +312,8 @@ namespace Bot.Commands {
                 };
 
                 var authoredLavalinkTracks = new List<AuthoredLavalinkTrack>();
-                if (i == -2) authoredLavalinkTracks.AddRange(tracks.Take(10).Select(track => AuthoredLavalinkTrack.FromLavalinkTrack(track, Context.User.Username)));
+                if (i == -2)
+                    authoredLavalinkTracks.AddRange(tracks.Take(10).Select(track => AuthoredLavalinkTrack.FromLavalinkTrack(track, Context.User.Username)));
                 if (i >= 0 && i <= tracks.Count - 1) authoredLavalinkTracks.Add(AuthoredLavalinkTrack.FromLavalinkTrack(tracks[i], Context.User.Username));
                 switch (authoredLavalinkTracks.Count) {
                     case 0:
@@ -378,39 +354,59 @@ namespace Bot.Commands {
         [Command("fastforward", RunMode = RunMode.Async)]
         [Alias("fast forward", "ff", "fwd")]
         [Summary("fastforward0s")]
-        public async Task FastForward([Summary("fastforward0_0s")]TimeSpan? timeSpan = null) {
+        public async Task FastForward([Summary("fastforward0_0s")] TimeSpan? timeSpan = null) {
             if (!await IsPreconditionsValid) return;
+            if (Player?.CurrentTrack == null) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
+                return;
+            }
+            
             var time = timeSpan ?? new TimeSpan(0, 0, 10);
             Player.SeekPositionAsync(Player.TrackPosition + time);
             Player.WriteToQueueHistory(Loc.Get("MusicQueues.FF").Format(Context.User.Username, Player.CurrentTrackIndex, time.TotalSeconds));
         }
-        
+
         [Command("rewind", RunMode = RunMode.Async)]
         [Alias("rw")]
         [Summary("rewind0s")]
-        public async Task Rewind([Summary("fastforward0_0s")]TimeSpan? timeSpan = null) {
+        public async Task Rewind([Summary("fastforward0_0s")] TimeSpan? timeSpan = null) {
             if (!await IsPreconditionsValid) return;
+            if (Player?.CurrentTrack == null) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
+                return;
+            }
+            
             var time = timeSpan ?? new TimeSpan(0, 0, 10);
             Player.SeekPositionAsync(Player.TrackPosition - time);
             Player.WriteToQueueHistory(Loc.Get("MusicQueues.Rewind").Format(Context.User.Username, Player.CurrentTrackIndex, time.TotalSeconds));
         }
-        
+
         [Command("seek", RunMode = RunMode.Async)]
         [Alias("sk", "se")]
         [Summary("seek0s")]
-        public async Task Seek([Summary("seek0_0s")]TimeSpan position) {
+        public async Task Seek([Summary("seek0_0s")] TimeSpan position) {
             if (!await IsPreconditionsValid) return;
+            if (Player?.CurrentTrack == null) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
+                return;
+            }
+            
             Player.SeekPositionAsync(position);
             Player.WriteToQueueHistory(Loc.Get("MusicQueues.Seek").Format(Context.User.Username, Player.CurrentTrackIndex, position));
         }
-        
+
         [Command("removerange", RunMode = RunMode.Async)]
         [Alias("rr", "delr", "dr")]
         [Summary("remove0s")]
-        public async Task RemoveRange([Summary("remove0_0s")]int start, [Summary("remove0_1s")]int end = -1) {
+        public async Task RemoveRange([Summary("remove0_0s")] int start, [Summary("remove0_1s")] int end = -1) {
             if (!await IsPreconditionsValid) return;
-            start = Math.Max(1, Math.Min(start, Player.Playlist.Count));
-            end = Math.Max(start, Math.Min(end, Player.Playlist.Count));
+            if (Player == null || Player.Playlist.IsEmpty) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
+                return;
+            }
+
+            start = start.Normalize(1, Player.Playlist.Count);
+            end = end.Normalize(start, Player.Playlist.Count);
             var removesCurrent = Player.CurrentTrackIndex + 1 >= start && Player.CurrentTrackIndex + 1 <= end;
             Player.Playlist.RemoveRange(start - 1, end - start + 1);
             Player.WriteToQueueHistory(Loc.Get("MusicQueues.Remove").Format(Context.User.Username, end - start + 1, start, end));
@@ -418,19 +414,23 @@ namespace Bot.Commands {
                 Jump();
             }
         }
-        
+
         [Command("remove", RunMode = RunMode.Async)]
         [Alias("rm", "del", "delete")]
         [Summary("remove0s")]
-        public async Task Remove([Summary("remove0_0s")]int start, [Summary("remove1_1s")]int count = 1) {
+        public async Task Remove([Summary("remove0_0s")] int start, [Summary("remove1_1s")] int count = 1) {
             await RemoveRange(start, start + count - 1);
         }
-        
+
         [Command("move", RunMode = RunMode.Async)]
         [Alias("m", "mv")]
         [Summary("move0s")]
-        public async Task Move([Summary("move0_0s")]int trackIndex, [Summary("move0_1s")]int newIndex = 1) {
+        public async Task Move([Summary("move0_0s")] int trackIndex, [Summary("move0_1s")] int newIndex = 1) {
             if (!await IsPreconditionsValid) return;
+            if (Player == null || Player.Playlist.IsEmpty) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
+                return;
+            }
             
             // For programmers
             if (trackIndex == 0) trackIndex = 1;
@@ -438,10 +438,77 @@ namespace Bot.Commands {
                 ReplyFormattedAsync(Loc.Get("Music.TrackIndexWrong").Format(Context.User.Mention, trackIndex, Player.Playlist.Count),
                     true).DelayedDelete(TimeSpan.FromMinutes(2));
             }
-            
+
             newIndex = Math.Max(1, Math.Min(Player.Playlist.Count, newIndex));
             Player.Playlist.Move(trackIndex - 1, newIndex - 1);
             Player.WriteToQueueHistory(Loc.Get("MusicQueues.TrackMoved").Format(Context.User.Mention, trackIndex, newIndex));
+        }
+
+        [Command("bassboost", RunMode = RunMode.Async)]
+        [Alias("bb", "bassboosted")]
+        [Summary("bassboost0s")]
+        public async Task ApplyBassBoost([Summary("bassboost0_0s")] BassBoostMode mode) {
+            if (!await IsPreconditionsValid) return;
+            if (Player == null) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
+                return;
+            }
+            
+            var bands = new List<EqualizerBand>();
+            switch (mode) {
+                case BassBoostMode.Off:
+                    bands.Add(new EqualizerBand(0, 0f));
+                    bands.Add(new EqualizerBand(1, 0f));
+                    break;
+                case BassBoostMode.Low:
+                    bands.Add(new EqualizerBand(0, 0.25f));
+                    bands.Add(new EqualizerBand(1, 0.15f));
+                    break;
+                case BassBoostMode.Medium:
+                    bands.Add(new EqualizerBand(0, 0.5f));
+                    bands.Add(new EqualizerBand(1, 0.25f));
+                    break;
+                case BassBoostMode.High:
+                    bands.Add(new EqualizerBand(0, 0.75f));
+                    bands.Add(new EqualizerBand(1, 0.5f));
+                    break;
+                case BassBoostMode.Extreme:
+                    bands.Add(new EqualizerBand(0, 1f));
+                    bands.Add(new EqualizerBand(1, 0.75f));
+                    break;
+            }
+
+            await Player.UpdateEqualizerAsync(bands, false);
+            Player.BassBoostMode = mode;
+            Player.UpdateParameters();
+            Player.WriteToQueueHistory(Loc.Get("MusicQueues.BassBoostUpdated").Format(Context.User.Username, mode));
+        }
+
+        [Command("changenode", RunMode = RunMode.Async)]
+        [Alias("newnode", "switchnode")]
+        [Summary("changenode0s")]
+        public async Task ChangeNode() {
+            if (!await IsPreconditionsValid) return;
+            if (Player == null) {
+                ReplyFormattedAsync(Loc.Get("Music.NothingPlaying").Format(GuildConfig.Prefix), true).DelayedDelete(TimeSpan.FromMinutes(1));
+                return;
+            }
+
+            if (MusicUtils.Cluster.Nodes.Count <= 1) {
+                ReplyFormattedAsync(Loc.Get("Music.OnlyOneNode"), true);
+                return;
+            }
+
+            var currentNode = MusicUtils.Cluster.GetServingNode(Context.Guild.Id);
+            var newNode = MusicUtils.Cluster.Nodes.Where(node => node.IsConnected).Where(node => node != currentNode).RandomOrDefault();
+            if (newNode == null) {
+                ReplyFormattedAsync(Loc.Get("Music.OnlyOneNode"), true);
+                return;
+            }
+
+            await currentNode.MovePlayerAsync(Player, newNode);
+            Player.WriteToQueueHistory(Loc.Get("MusicQueues.NodeChanged").Format(Context.User.Username, newNode.Label));
+            Player.UpdateNodeName();
         }
     }
 }
